requirements.txt
tensorflow==2.10.0
librosa==0.9.2
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
sounddevice==0.4.6
matplotlib==3.7.1
websockets==11.0.3
ffmpeg-python==0.2.0
kagglehub==0.1.0